+++ src/realm/views.cpp
@@ -21,6 +21,8 @@
 #include <realm/column_link.hpp>
 #include <realm/table.hpp>
 
+#include <typeinfo>
+
 using namespace realm;
 
 namespace {
@@ -30,7 +32,7 @@ struct IndexPair {
 };
 } // anonymous namespace
 
-CommonDescriptor::CommonDescriptor(Table const& table, std::vector<std::vector<size_t>> column_indices)
+ColumnsDescriptor::ColumnsDescriptor(Table const& table, std::vector<std::vector<size_t>> column_indices)
 {
     if (table.is_degenerate()) {
         // We need access to the column acessors and that's not available in a
@@ -60,14 +62,14 @@ CommonDescriptor::CommonDescriptor(Table const& table, std::vector<std::vector<s
     }
 }
 
-std::unique_ptr<CommonDescriptor> CommonDescriptor::clone() const
+std::unique_ptr<BaseDescriptor> ColumnsDescriptor::clone() const
 {
-    return std::unique_ptr<CommonDescriptor>(new CommonDescriptor(*this));
+    return std::unique_ptr<BaseDescriptor>(new ColumnsDescriptor(*this));
 }
 
 SortDescriptor::SortDescriptor(Table const& table, std::vector<std::vector<size_t>> column_indices,
                                std::vector<bool> ascending)
-    : CommonDescriptor(table, column_indices)
+    : ColumnsDescriptor(table, column_indices)
     , m_ascending(std::move(ascending))
 {
     REALM_ASSERT_EX(m_ascending.empty() || m_ascending.size() == column_indices.size(), m_ascending.size(),
@@ -79,9 +81,9 @@ SortDescriptor::SortDescriptor(Table const& table, std::vector<std::vector<size_
     }
 }
 
-std::unique_ptr<CommonDescriptor> SortDescriptor::clone() const
+std::unique_ptr<BaseDescriptor> SortDescriptor::clone() const
 {
-    return std::unique_ptr<CommonDescriptor>(new SortDescriptor(*this));
+    return std::unique_ptr<ColumnsDescriptor>(new SortDescriptor(*this));
 }
 
 void SortDescriptor::merge_with(SortDescriptor&& other)
@@ -96,7 +98,7 @@ void SortDescriptor::merge_with(SortDescriptor&& other)
                        other.m_ascending.end());
 }
 
-class CommonDescriptor::Sorter {
+class ColumnsDescriptor::Sorter {
 public:
     Sorter(std::vector<std::vector<const ColumnBase*>> const& columns, std::vector<bool> const& ascending,
            IntegerColumn const& row_indexes);
@@ -126,7 +128,7 @@ private:
     std::vector<SortColumn> m_columns;
 };
 
-CommonDescriptor::Sorter::Sorter(std::vector<std::vector<const ColumnBase*>> const& columns,
+ColumnsDescriptor::Sorter::Sorter(std::vector<std::vector<const ColumnBase*>> const& columns,
                                  std::vector<bool> const& ascending, IntegerColumn const& row_indexes)
 {
     REALM_ASSERT(!columns.empty());
@@ -149,7 +151,7 @@ CommonDescriptor::Sorter::Sorter(std::vector<std::vector<const ColumnBase*>> con
         for (size_t row_ndx = 0; row_ndx < num_rows; row_ndx++) {
             size_t translated_index = to_size_t(row_indexes.get(row_ndx));
             for (size_t j = 0; j + 1 < columns[i].size(); ++j) {
-                // type was checked when creating the CommonDescriptor
+                // type was checked when creating the ColumnsDescriptor
                 auto link_col = static_cast<const LinkColumn*>(columns[i][j]);
                 if (link_col->is_null(translated_index)) {
                     is_null[row_ndx] = true;
@@ -162,7 +164,7 @@ CommonDescriptor::Sorter::Sorter(std::vector<std::vector<const ColumnBase*>> con
     }
 }
 
-std::vector<std::vector<size_t>> CommonDescriptor::export_column_indices() const
+DescriptorExport ColumnsDescriptor::export_for_handover() const
 {
     std::vector<std::vector<size_t>> column_indices;
     column_indices.reserve(m_columns.size());
@@ -173,12 +175,18 @@ std::vector<std::vector<size_t>> CommonDescriptor::export_column_indices() const
             indices.push_back(col->get_column_index());
         column_indices.push_back(indices);
     }
-    return column_indices;
+    DescriptorExport out;
+    out.type = DescriptorType::Distinct;
+    out.columns = column_indices;
+    return out;
 }
 
-std::vector<bool> SortDescriptor::export_order() const
+DescriptorExport SortDescriptor::export_for_handover() const
 {
-    return m_ascending;
+    DescriptorExport out = ColumnsDescriptor::export_for_handover();
+    out.type = DescriptorType::Sort;
+    out.ordering = m_ascending;
+    return out;
 }
 
 std::string SortDescriptor::get_description(TableRef attached_table) const
@@ -213,7 +221,7 @@ std::string SortDescriptor::get_description(TableRef attached_table) const
     return description;
 }
 
-std::string CommonDescriptor::get_description(TableRef attached_table) const
+std::string ColumnsDescriptor::get_description(TableRef attached_table) const
 {
     std::string description = "DISTINCT(";
     for (size_t i = 0; i < m_columns.size(); ++i) {
@@ -237,7 +245,7 @@ std::string CommonDescriptor::get_description(TableRef attached_table) const
     return description;
 }
 
-CommonDescriptor::Sorter CommonDescriptor::sorter(IntegerColumn const& row_indexes) const
+ColumnsDescriptor::Sorter ColumnsDescriptor::sorter(IntegerColumn const& row_indexes) const
 {
     REALM_ASSERT(!m_columns.empty());
     std::vector<bool> ascending(m_columns.size(), true);
@@ -280,6 +288,29 @@ bool SortDescriptor::Sorter::operator()(IndexPair i, IndexPair j, bool total_ord
     return total_ordering ? i.index_in_view < j.index_in_view : 0;
 }
 
+LimitDescriptor::LimitDescriptor(size_t limit)
+    : m_limit(limit)
+{
+}
+
+std::string LimitDescriptor::get_description(TableRef) const
+{
+    return "LIMIT(" + serializer::print_value(m_limit) + ")";
+}
+
+std::unique_ptr<BaseDescriptor> LimitDescriptor::clone() const
+{
+    return std::unique_ptr<BaseDescriptor>(new LimitDescriptor(*this));
+}
+
+DescriptorExport LimitDescriptor::export_for_handover() const
+{
+    DescriptorExport out;
+    out.type = DescriptorType::Limit;
+    out.limit = m_limit;
+    return out;
+}
+
 DescriptorOrdering::DescriptorOrdering(const DescriptorOrdering& other)
 {
     for (const auto& d : other.m_descriptors) {
@@ -303,11 +334,10 @@ void DescriptorOrdering::append_sort(SortDescriptor sort)
     if (!sort.is_valid()) {
         return;
     }
-    if (!m_descriptors.empty()) {
-        if (SortDescriptor* previous_sort = dynamic_cast<SortDescriptor*>(m_descriptors.back().get())) {
-            previous_sort->merge_with(std::move(sort));
-            return;
-        }
+    if (!m_descriptors.empty() && m_descriptors.back()->get_type() == DescriptorType::Sort) {
+        SortDescriptor* previous_sort = static_cast<SortDescriptor*>(m_descriptors.back().get());
+        previous_sort->merge_with(std::move(sort));
+        return;
     }
     m_descriptors.emplace_back(new SortDescriptor(std::move(sort)));
 }
@@ -319,36 +349,84 @@ void DescriptorOrdering::append_distinct(DistinctDescriptor distinct)
     }
 }
 
+void DescriptorOrdering::append_limit(LimitDescriptor limit)
+{
+    REALM_ASSERT(limit.is_valid());
+    m_descriptors.emplace_back(new LimitDescriptor(std::move(limit)));
+}
+
 bool DescriptorOrdering::descriptor_is_sort(size_t index) const
 {
     REALM_ASSERT(index < m_descriptors.size());
-    SortDescriptor* sort_descr = dynamic_cast<SortDescriptor*>(m_descriptors[index].get());
-    return (sort_descr != nullptr);
+    return m_descriptors[index]->get_type() == DescriptorType::Sort;
 }
 
 bool DescriptorOrdering::descriptor_is_distinct(size_t index) const
 {
-    return !descriptor_is_sort(index);
+    REALM_ASSERT(index < m_descriptors.size());
+    return m_descriptors[index]->get_type() == DescriptorType::Distinct;
 }
 
-const CommonDescriptor* DescriptorOrdering::operator[](size_t ndx) const
+bool DescriptorOrdering::descriptor_is_limit(size_t index) const
+{
+    REALM_ASSERT(index < m_descriptors.size());
+    return m_descriptors[index]->get_type() == DescriptorType::Limit;
+}
+
+DescriptorType DescriptorOrdering::get_type(size_t index) const
+{
+    REALM_ASSERT(index < m_descriptors.size());
+    return m_descriptors[index]->get_type();
+}
+
+const BaseDescriptor* DescriptorOrdering::operator[](size_t ndx) const
 {
     return m_descriptors.at(ndx).get(); // may throw std::out_of_range
 }
 
 bool DescriptorOrdering::will_apply_sort() const
 {
-    return std::any_of(m_descriptors.begin(), m_descriptors. end(), [](const std::unique_ptr<CommonDescriptor>& desc) {
+    return std::any_of(m_descriptors.begin(), m_descriptors.end(), [](const std::unique_ptr<BaseDescriptor>& desc) {
         REALM_ASSERT(desc.get()->is_valid());
-        return dynamic_cast<SortDescriptor*>(desc.get()) != nullptr;
+        return desc->get_type() == DescriptorType::Sort;
     });
 }
 
 bool DescriptorOrdering::will_apply_distinct() const
 {
-    return std::any_of(m_descriptors.begin(), m_descriptors.end(), [](const std::unique_ptr<CommonDescriptor>& desc) {
+    return std::any_of(m_descriptors.begin(), m_descriptors.end(), [](const std::unique_ptr<BaseDescriptor>& desc) {
         REALM_ASSERT(desc.get()->is_valid());
-        return dynamic_cast<SortDescriptor*>(desc.get()) == nullptr;
+        return desc->get_type() == DescriptorType::Distinct;
+    });
+}
+
+bool DescriptorOrdering::will_apply_limit() const
+{
+    return std::any_of(m_descriptors.begin(), m_descriptors.end(), [](const std::unique_ptr<BaseDescriptor>& desc) {
+        REALM_ASSERT(desc.get()->is_valid());
+        return desc->get_type() == DescriptorType::Limit;
+    });
+}
+
+realm::util::Optional<size_t> DescriptorOrdering::get_min_limit() const
+{
+    realm::util::Optional<size_t> min_limit;
+    for (size_t i = 0; i < m_descriptors.size(); ++i) {
+        if (m_descriptors[i]->get_type() == DescriptorType::Limit) {
+            const LimitDescriptor* limit = static_cast<const LimitDescriptor*>(m_descriptors[i].get());
+            REALM_ASSERT(limit);
+            min_limit = bool(min_limit) ? std::min(*min_limit, limit->get_limit()) : limit->get_limit();
+        }
+    }
+    return min_limit;
+}
+
+bool DescriptorOrdering::will_limit_to_zero() const
+{
+    return std::any_of(m_descriptors.begin(), m_descriptors.end(), [](const std::unique_ptr<BaseDescriptor>& desc) {
+        REALM_ASSERT(desc.get()->is_valid());
+        return (desc->get_type() == DescriptorType::Limit &&
+                static_cast<LimitDescriptor*>(desc.get())->get_limit() == 0);
     });
 }
 
@@ -369,16 +447,13 @@ void DescriptorOrdering::generate_patch(DescriptorOrdering const& descriptors, H
 {
     if (!descriptors.is_empty()) {
         const size_t num_descriptors = descriptors.size();
-        std::vector<std::vector<std::vector<size_t>>> column_indices;
-        std::vector<std::vector<bool>> column_orders;
-        column_indices.reserve(num_descriptors);
-        column_orders.reserve(num_descriptors);
+        std::vector<DescriptorExport> out;
+        out.reserve(num_descriptors);
         for (size_t desc_ndx = 0; desc_ndx < num_descriptors; ++desc_ndx) {
-            const CommonDescriptor* desc = descriptors[desc_ndx];
-            column_indices.push_back(desc->export_column_indices());
-            column_orders.push_back(desc->export_order());
+            const BaseDescriptor* desc = descriptors[desc_ndx];
+            out.push_back(desc->export_for_handover());
         }
-        patch.reset(new DescriptorOrderingHandoverPatch{std::move(column_indices), std::move(column_orders)});
+        patch.reset(new DescriptorOrderingHandoverPatch{std::move(out)});
     }
 }
 
@@ -386,17 +461,18 @@ DescriptorOrdering DescriptorOrdering::create_from_and_consume_patch(HandoverPat
 {
     DescriptorOrdering ordering;
     if (patch) {
-        const size_t num_descriptors = patch->columns.size();
-        REALM_ASSERT_EX(num_descriptors == patch->ascending.size(),
-                        num_descriptors, patch->ascending.size());
-        for (size_t desc_ndx = 0; desc_ndx < num_descriptors; ++desc_ndx) {
-            if (patch->columns[desc_ndx].size() != patch->ascending[desc_ndx].size()) {
-                // If size differs, it must be a distinct
-                ordering.append_distinct(DistinctDescriptor(table, std::move(patch->columns[desc_ndx])));
-            }
-            else {
-                ordering.append_sort(SortDescriptor(table, std::move(patch->columns[desc_ndx]),
-                                                    std::move(patch->ascending[desc_ndx])));
+        for (size_t desc_ndx = 0; desc_ndx < patch->descriptors.size(); ++desc_ndx) {
+            DescriptorExport single = patch->descriptors[desc_ndx];
+            switch (single.type) {
+                case DescriptorType::Sort:
+                    ordering.append_sort(SortDescriptor(table, std::move(single.columns), std::move(single.ordering)));
+                    break;
+                case DescriptorType::Distinct:
+                    ordering.append_distinct(DistinctDescriptor(table, std::move(single.columns)));
+                    break;
+                case DescriptorType::Limit:
+                    ordering.append_limit(LimitDescriptor(single.limit));
+                    break;
             }
         }
         patch.reset();
@@ -405,8 +481,10 @@ DescriptorOrdering DescriptorOrdering::create_from_and_consume_patch(HandoverPat
 }
 
 void RowIndexes::do_sort(const DescriptorOrdering& ordering) {
+    m_limit_count = 0;
     if (ordering.is_empty())
         return;
+
     size_t sz = size();
     if (sz == 0)
         return;
@@ -430,60 +508,80 @@ void RowIndexes::do_sort(const DescriptorOrdering& ordering) {
 
     const int num_descriptors = int(ordering.size());
     for (int desc_ndx = 0; desc_ndx < num_descriptors; ++desc_ndx) {
-        const CommonDescriptor* common_descr = ordering[desc_ndx];
 
-        if (const auto* sort_descr = dynamic_cast<const SortDescriptor*>(common_descr)) {
+        DescriptorType type = ordering.get_type(desc_ndx);
+        switch (type) {
+            case DescriptorType::Sort:
+            {
+                const auto* sort_descr = static_cast<const SortDescriptor*>(ordering[desc_ndx]);
+                SortDescriptor::Sorter sort_predicate = sort_descr->sorter(m_row_indexes);
 
-            SortDescriptor::Sorter sort_predicate = sort_descr->sorter(m_row_indexes);
+                std::sort(v.begin(), v.end(), std::ref(sort_predicate));
 
-            std::sort(v.begin(), v.end(), std::ref(sort_predicate));
-
-            bool is_last_ordering = desc_ndx == num_descriptors - 1;
-            // not doing this on the last step is an optimisation
-            if (!is_last_ordering) {
-                const size_t v_size = v.size();
-                // Distinct must choose the winning unique elements by sorted
-                // order not by the previous tableview order, the lowest
-                // "index_in_view" wins.
-                for (size_t i = 0; i < v_size; ++i) {
-                    v[i].index_in_view = i;
+                bool is_last_ordering = desc_ndx == num_descriptors - 1;
+                // not doing this on the last step is an optimisation
+                if (!is_last_ordering) {
+                    const size_t v_size = v.size();
+                    // Distinct must choose the winning unique elements by sorted
+                    // order not by the previous tableview order, the lowest
+                    // "index_in_view" wins.
+                    for (size_t i = 0; i < v_size; ++i) {
+                        v[i].index_in_view = i;
+                    }
                 }
+                break;
             }
-        }
-        else { // distinct descriptor
-            auto distinct_predicate = common_descr->sorter(m_row_indexes);
+            case DescriptorType::Distinct:
+            {
+                const auto* distinct_descr = static_cast<const DistinctDescriptor*>(ordering[desc_ndx]);
+                auto distinct_predicate = distinct_descr->sorter(m_row_indexes);
 
-            // Remove all rows which have a null link along the way to the distinct columns
-            if (distinct_predicate.has_links()) {
-                v.erase(std::remove_if(v.begin(), v.end(),
-                                       [&](auto&& index) { return distinct_predicate.any_is_null(index); }),
+                // Remove all rows which have a null link along the way to the distinct columns
+                if (distinct_predicate.has_links()) {
+                    v.erase(std::remove_if(v.begin(), v.end(),
+                                           [&](auto&& index) { return distinct_predicate.any_is_null(index); }),
+                            v.end());
+                }
+
+                // Sort by the columns to distinct on
+                std::sort(v.begin(), v.end(), std::ref(distinct_predicate));
+
+                // Remove all duplicates
+                v.erase(std::unique(v.begin(), v.end(),
+                                    [&](auto&& a, auto&& b) {
+                                        // "not less than" is "equal" since they're sorted
+                                        return !distinct_predicate(a, b, false);
+                                    }),
                         v.end());
+                bool will_be_sorted_next = desc_ndx < num_descriptors - 1 && ordering.descriptor_is_sort(desc_ndx + 1);
+                if (!will_be_sorted_next) {
+                    // Restore the original order, this is either the original
+                    // tableview order or the order of the previous sort
+                    std::sort(v.begin(), v.end(), [](auto a, auto b) { return a.index_in_view < b.index_in_view; });
+                }
+
+                break;
             }
-
-            // Sort by the columns to distinct on
-            std::sort(v.begin(), v.end(), std::ref(distinct_predicate));
-
-            // Remove all duplicates
-            v.erase(std::unique(v.begin(), v.end(),
-                                [&](auto&& a, auto&& b) {
-                                    // "not less than" is "equal" since they're sorted
-                                    return !distinct_predicate(a, b, false);
-                                }),
-                    v.end());
-            bool will_be_sorted_next = desc_ndx < num_descriptors - 1 && ordering.descriptor_is_sort(desc_ndx + 1);
-            if (!will_be_sorted_next) {
-                // Restore the original order, this is either the original
-                // tableview order or the order of the previous sort
-                std::sort(v.begin(), v.end(), [](auto a, auto b) { return a.index_in_view < b.index_in_view; });
+            case DescriptorType::Limit:
+            {
+                const auto* limit_descr = static_cast<const LimitDescriptor*>(ordering[desc_ndx]);
+                const size_t limit = limit_descr->get_limit();
+                if (v.size() > limit) {
+                    m_limit_count += v.size() - limit;
+                    v.erase(v.begin() + limit, v.end());
+                }
+                break;
             }
         }
     }
     // Apply the results
     m_row_indexes.clear();
-    for (auto& pair : v)
+    for (auto& pair : v) {
         m_row_indexes.add(pair.index_in_column);
-    for (size_t t = 0; t < detached_ref_count; ++t)
+    }
+    for (size_t t = 0; t < detached_ref_count; ++t) {
         m_row_indexes.add(-1);
+    }
 }
 
 RowIndexes::RowIndexes(IntegerColumn::unattached_root_tag urt, realm::Allocator& alloc)
@@ -506,8 +604,9 @@ RowIndexes::RowIndexes(IntegerColumn&& col)
 // FIXME: this only works (and is only used) for row indexes with memory
 // managed by the default allocator, e.q. for TableViews.
 RowIndexes::RowIndexes(const RowIndexes& source, ConstSourcePayload mode)
+    : m_limit_count(source.m_limit_count)
 #ifdef REALM_COOKIE_CHECK
-    : m_debug_cookie(source.m_debug_cookie)
+    , m_debug_cookie(source.m_debug_cookie)
 #endif
 {
     REALM_ASSERT(&source.m_row_indexes.get_alloc() == &Allocator::get_default());
@@ -519,8 +618,9 @@ RowIndexes::RowIndexes(const RowIndexes& source, ConstSourcePayload mode)
 }
 
 RowIndexes::RowIndexes(RowIndexes& source, MutableSourcePayload)
+    : m_limit_count(source.m_limit_count)
 #ifdef REALM_COOKIE_CHECK
-    : m_debug_cookie(source.m_debug_cookie)
+    , m_debug_cookie(source.m_debug_cookie)
 #endif
 {
     REALM_ASSERT(&source.m_row_indexes.get_alloc() == &Allocator::get_default());