+++ test/test_parser.cpp
@@ -197,6 +197,19 @@ static std::vector<std::string> valid_queries = {
     "a == b and c==d sort(a ASC, b DESC) DISTINCT(p) sort(c ASC, d DESC) DISTINCT(q.r)",
     "a == b  sort(     a   ASC  ,  b DESC) and c==d   DISTINCT(   p )  sort(   c   ASC  ,  d   DESC  )  DISTINCT(   q.r ,   p)   ",
 
+    // limit
+    "a=b LIMIT(1)",
+    "a=b LIMIT ( 1 )",
+    "a=b LIMIT( 1234567890 )",
+    "a=b LIMIT(1) && c=d",
+    "a=b && c=d || e=f LIMIT(1)",
+    "a=b LIMIT(1) SORT(a ASC) DISTINCT(b)",
+    "a=b SORT(a ASC) LIMIT(1) DISTINCT(b)",
+    "a=b SORT(a ASC) DISTINCT(b) LIMIT(1)",
+    "a=b LIMIT(2) LIMIT(1)",
+    "a=b LIMIT(5) && c=d LIMIT(2)",
+    "a=b LIMIT(5) SORT(age ASC) DISTINCT(name) LIMIT(2)",
+
     // subquery expression
     "SUBQUERY(items, $x, $x.name == 'Tom').@size > 0",
     "SUBQUERY(items, $x, $x.name == 'Tom').@count > 0",
@@ -279,6 +292,19 @@ static std::vector<std::string> invalid_queries = {
     "a=b DISTINCT(p", // no braces
     "a=b sort(p.q DESC a ASC)", // missing comma
     "a=b DISTINCT(p q)", // missing comma
+
+    // limit
+    "LIMIT(1)", // no query conditions
+    "a=b LIMIT", // no params
+    "a=b LIMIT()", // no params
+    "a=b LIMIT(2", // missing end paren
+    "a=b LIMIT2)", // missing open paren
+    "a=b LIMIT(-1)", // negative limit
+    "a=b LIMIT(2.7)", // input must be an integer
+    "a=b LIMIT(0xFFEE)", // input must be an integer
+    "a=b LIMIT(word)", // non numeric limit
+    "a=b LIMIT(11asdf)", // non numeric limit
+    "a=b LIMIT(1, 1)", // only accept one input
 
     // subquery
     "SUBQUERY(items, $x, $x.name == 'Tom') > 0", // missing .@count
@@ -1616,7 +1642,6 @@ TableView get_sorted_view(TableRef t, std::string query_string)
     realm::query_builder::apply_predicate(q, result.predicate, args);
     DescriptorOrdering ordering;
     realm::query_builder::apply_ordering(ordering, t, result.ordering);
-
     std::string query_description = q.get_description();
     std::string ordering_description = ordering.get_description(t);
     std::string combined = query_description + " " + ordering_description;
@@ -1740,6 +1765,217 @@ TEST(Parser_SortAndDistinct)
 
     CHECK_THROW_ANY_GET_MESSAGE(get_sorted_view(people, "TRUEPREDICATE sort(account.name ASC)"), message);
     CHECK_EQUAL(message, "No property 'name' found on object type 'account' specified in 'sort' clause");
+}
+
+
+TEST(Parser_Limit)
+{
+    SHARED_GROUP_TEST_PATH(path);
+    std::unique_ptr<Replication> hist(make_in_realm_history(path));
+    SharedGroup sg(*hist, SharedGroupOptions(crypt_key()));
+
+    Group& g = sg.begin_write();
+    TableRef people = g.add_table("person");
+
+    size_t name_col = people->add_column(type_String, "name");
+    size_t age_col = people->add_column(type_Int, "age");
+
+    people->add_empty_row(3);
+    people->set_string(name_col, 0, "Adam");
+    people->set_int(age_col, 0, 28);
+    people->set_string(name_col, 1, "Frank");
+    people->set_int(age_col, 1, 30);
+    people->set_string(name_col, 2, "Ben");
+    people->set_int(age_col, 2, 28);
+
+    // solely limit
+    TableView tv = get_sorted_view(people, "TRUEPREDICATE LIMIT(0)");
+    CHECK_EQUAL(tv.size(), 0);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 3);
+    tv = get_sorted_view(people, "TRUEPREDICATE LIMIT(1)");
+    CHECK_EQUAL(tv.size(), 1);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 2);
+    tv = get_sorted_view(people, "TRUEPREDICATE LIMIT(2)");
+    CHECK_EQUAL(tv.size(), 2);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 1);
+    tv = get_sorted_view(people, "TRUEPREDICATE LIMIT(3)");
+    CHECK_EQUAL(tv.size(), 3);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    tv = get_sorted_view(people, "TRUEPREDICATE LIMIT(4)");
+    CHECK_EQUAL(tv.size(), 3);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+
+    // sort + limit
+    tv = get_sorted_view(people, "TRUEPREDICATE SORT(name ASC) LIMIT(0)");
+    CHECK_EQUAL(tv.size(), 0);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 3);
+    tv = get_sorted_view(people, "TRUEPREDICATE SORT(name ASC) LIMIT(1)");
+    CHECK_EQUAL(tv.size(), 1);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 2);
+    CHECK_EQUAL(tv.get_string(name_col, 0), "Adam");
+    tv = get_sorted_view(people, "TRUEPREDICATE SORT(name ASC) LIMIT(2)");
+    CHECK_EQUAL(tv.size(), 2);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 1);
+    CHECK_EQUAL(tv.get_string(name_col, 0), "Adam");
+    CHECK_EQUAL(tv.get_string(name_col, 1), "Ben");
+    tv = get_sorted_view(people, "TRUEPREDICATE SORT(name ASC) LIMIT(3)");
+    CHECK_EQUAL(tv.size(), 3);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    CHECK_EQUAL(tv.get_string(name_col, 0), "Adam");
+    CHECK_EQUAL(tv.get_string(name_col, 1), "Ben");
+    CHECK_EQUAL(tv.get_string(name_col, 2), "Frank");
+    tv = get_sorted_view(people, "TRUEPREDICATE SORT(name ASC) LIMIT(4)");
+    CHECK_EQUAL(tv.size(), 3);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+
+    // sort + distinct + limit
+    tv = get_sorted_view(people, "TRUEPREDICATE SORT(name ASC) DISTINCT(age) LIMIT(0)");
+    CHECK_EQUAL(tv.size(), 0);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 2);
+    tv = get_sorted_view(people, "TRUEPREDICATE SORT(name ASC) DISTINCT(age) LIMIT(1)");
+    CHECK_EQUAL(tv.size(), 1);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 1);
+    CHECK_EQUAL(tv.get_string(name_col, 0), "Adam");
+    tv = get_sorted_view(people, "TRUEPREDICATE SORT(name ASC) DISTINCT(age) LIMIT(2)");
+    CHECK_EQUAL(tv.size(), 2);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    CHECK_EQUAL(tv.get_string(name_col, 0), "Adam");
+    CHECK_EQUAL(tv.get_string(name_col, 1), "Frank");
+    tv = get_sorted_view(people, "TRUEPREDICATE SORT(name ASC) DISTINCT(age) LIMIT(3)");
+    CHECK_EQUAL(tv.size(), 2);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    CHECK_EQUAL(tv.get_string(name_col, 0), "Adam");
+    CHECK_EQUAL(tv.get_string(name_col, 1), "Frank");
+    tv = get_sorted_view(people, "TRUEPREDICATE SORT(name ASC) DISTINCT(age) LIMIT(4)");
+    CHECK_EQUAL(tv.size(), 2);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+
+    // query + limit
+    tv = get_sorted_view(people, "age < 30 SORT(name ASC) DISTINCT(age) LIMIT(0)");
+    CHECK_EQUAL(tv.size(), 0);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 1);
+    tv = get_sorted_view(people, "age < 30 SORT(name ASC) DISTINCT(age) LIMIT(1)");
+    CHECK_EQUAL(tv.size(), 1);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    CHECK_EQUAL(tv.get_string(name_col, 0), "Adam");
+    tv = get_sorted_view(people, "age < 30 SORT(name ASC) DISTINCT(age) LIMIT(2)");
+    CHECK_EQUAL(tv.size(), 1);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    CHECK_EQUAL(tv.get_string(name_col, 0), "Adam");
+    tv = get_sorted_view(people, "age < 30 SORT(name ASC) DISTINCT(age) LIMIT(3)");
+    CHECK_EQUAL(tv.size(), 1);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    CHECK_EQUAL(tv.get_string(name_col, 0), "Adam");
+    tv = get_sorted_view(people, "age < 30 SORT(name ASC) DISTINCT(age) LIMIT(4)");
+    CHECK_EQUAL(tv.size(), 1);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+
+    // compound query + limit
+    tv = get_sorted_view(people, "age < 30 && name == 'Adam' LIMIT(0)");
+    CHECK_EQUAL(tv.size(), 0);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 1);
+    tv = get_sorted_view(people, "age < 30 && name == 'Adam' LIMIT(1)");
+    CHECK_EQUAL(tv.size(), 1);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    CHECK_EQUAL(tv.get_string(name_col, 0), "Adam");
+
+    // limit multiple times, order matters
+    tv = get_sorted_view(people, "TRUEPREDICATE LIMIT(2) LIMIT(1)");
+    CHECK_EQUAL(tv.size(), 1);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 2);
+    tv = get_sorted_view(people, "TRUEPREDICATE LIMIT(3) LIMIT(2) LIMIT(1) LIMIT(10)");
+    CHECK_EQUAL(tv.size(), 1);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 2);
+    tv = get_sorted_view(people, "age > 0 SORT(name ASC) LIMIT(2)");
+    CHECK_EQUAL(tv.size(), 2);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 1);
+    CHECK_EQUAL(tv.get_string(name_col, 0), "Adam");
+    CHECK_EQUAL(tv.get_string(name_col, 1), "Ben");
+    tv = get_sorted_view(people, "age > 0 LIMIT(2) SORT(name ASC)");
+    CHECK_EQUAL(tv.size(), 2);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 1);
+    CHECK_EQUAL(tv.get_string(name_col, 0), "Adam");
+    CHECK_EQUAL(tv.get_string(name_col, 1), "Frank");
+    tv = get_sorted_view(people, "age > 0 SORT(name ASC) LIMIT(2) DISTINCT(age)");
+    CHECK_EQUAL(tv.size(), 1);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 1); // the other result is excluded by distinct not limit
+    tv = get_sorted_view(people, "age > 0 SORT(name DESC) LIMIT(2) SORT(age ASC) LIMIT(1)");
+    CHECK_EQUAL(tv.size(), 1);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 2);
+    CHECK_EQUAL(tv.get_string(name_col, 0), "Ben");
+
+    // size_unlimited() checks
+    tv = get_sorted_view(people, "age == 30");
+    CHECK_EQUAL(tv.size(), 1);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    tv = get_sorted_view(people, "age == 30 LIMIT(0)");
+    CHECK_EQUAL(tv.size(), 0);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 1);
+    tv = get_sorted_view(people, "age == 1000");
+    CHECK_EQUAL(tv.size(), 0);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    tv = get_sorted_view(people, "age == 1000 LIMIT(0)");
+    CHECK_EQUAL(tv.size(), 0);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    tv = get_sorted_view(people, "age == 1000 SORT(name ASC)");
+    CHECK_EQUAL(tv.size(), 0);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    tv = get_sorted_view(people, "age == 1000 SORT(name ASC) LIMIT(0)");
+    CHECK_EQUAL(tv.size(), 0);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    tv = get_sorted_view(people, "age == 28 SORT(name ASC)");
+    CHECK_EQUAL(tv.size(), 2);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    tv = get_sorted_view(people, "age == 28 SORT(name ASC) LIMIT(1)");
+    CHECK_EQUAL(tv.size(), 1);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 1);
+    tv = get_sorted_view(people, "age == 28 DISTINCT(age)");
+    CHECK_EQUAL(tv.size(), 1);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    tv = get_sorted_view(people, "age == 28 DISTINCT(age) LIMIT(0)");
+    CHECK_EQUAL(tv.size(), 0);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 1);
+    tv = get_sorted_view(people, "age == 28 SORT(name ASC) DISTINCT(age)");
+    CHECK_EQUAL(tv.size(), 1);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    tv = get_sorted_view(people, "age == 28 SORT(name ASC) DISTINCT(age) LIMIT(0)");
+    CHECK_EQUAL(tv.size(), 0);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 1);
+    tv = get_sorted_view(people, "FALSEPREDICATE");
+    CHECK_EQUAL(tv.size(), 0);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    tv = get_sorted_view(people, "FALSEPREDICATE LIMIT(0)");
+    CHECK_EQUAL(tv.size(), 0);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+    tv = get_sorted_view(people, "FALSEPREDICATE LIMIT(1)");
+    CHECK_EQUAL(tv.size(), 0);
+    CHECK_EQUAL(tv.get_num_results_excluded_by_limit(), 0);
+
+    // errors
+    CHECK_THROW_ANY(get_sorted_view(people, "TRUEPREDICATE LIMIT(-1)")); // only accepting positive integers
+    CHECK_THROW_ANY(get_sorted_view(people, "TRUEPREDICATE LIMIT(age)")); // only accepting positive integers
+    CHECK_THROW_ANY(get_sorted_view(people, "TRUEPREDICATE LIMIT('age')")); // only accepting positive integers
+
+    sg.commit();
+
+    // handover
+    const Group& reader = sg.begin_read();
+    ConstTableRef peopleRead = reader.get_table("person");
+
+    TableView items = peopleRead->where().find_all();
+    CHECK_EQUAL(items.size(), 3);
+    realm::DescriptorOrdering desc;
+    CHECK(!desc.will_apply_limit());
+    desc.append_limit(1);
+    CHECK(desc.will_apply_limit());
+    items.apply_descriptor_ordering(desc);
+    CHECK_EQUAL(items.size(), 1);
+    std::unique_ptr<SharedGroup::Handover<TableView>> handover = sg.export_for_handover(items, ConstSourcePayload::Copy);
+
+    std::unique_ptr<TableView> tv2(sg.import_from_handover(move(handover)));
+    CHECK(tv2->is_attached());
+    CHECK(tv2->is_in_sync());
+    CHECK_EQUAL(tv2->size(), 1);
 }