+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 
 ### Enhancements
 
-* None.
+* Support limiting queries via `DescriptorOrdering::set_limit` and by supporting "LIMIT(x)" in string queries.
 
 -----------